**DISRUPTIVE CHANGE** begin refactor: move model code to bunsen/model.py
This starts to set up bunsen as a 'proper' module
(see branch module-design for a picture of the end goal) with core classes
split among several python files in subdirectory bunsen/.

The code passes basic sanity tests, but there may be import-breakage to fix later.
If in doubt, revert to tag pre-2021-refactor.