Add support for reading the target_board
The model has some minimal support for DejaGNU target_board already,
but it is not being recorded anywhere. This patch adds these missing
bits.

GDB testing regularly includes several target boards, e.g.,
unix/-m32, native-extended-gdbserver/-m64, and so on. The target
board is output to gdb.{log,sum} with the line

   Running target TARGET_BOARD

Note that Bunsen can currently only handle one target board at a time.
This is not an attempt to fix that yet.

After this patch, +summarize should be able to output this information
(instead of the current "<unknown board>". To use this feature,
data must obviously be re-imported.