Add new outcomes to summarize script
GDB has a number of custom test outcomes which occasionally appear
in test results. This patch adds the appropriate DejaGNU output string
for these outcomes.

This patch also outputs a footnote whenever the DUPLICATE outcome
is seen. For GDB, this number will not be the same as the number
reported in gdb.sum, since that reported number is the number
of unique duplicates. The number we record in Bunsen is the
total number of tests that have the DUPLICATE outcome.

$ grep "of duplicate test names" gdb.sum
# of duplicate test names	357

vs

$ grep ^DUPLICATE gdb.log | wc -l
428

It is this last number that the script reports.