bugfix Workdir.commit_all: don't default to allow_duplicates=False
This option is already provided explicitly by the code in
Bunsen.commit() that needs to use it to commit to testlogs branches
while avoiding duplicate creation.

Elsewhere, the option should default to True.  Otherwise,
functionality such as 'bunsen rm' will end up not making necessary
commits when they produce a previously-existing tree in one of the
index branches.